GameBeak#: Change of memoryPointer/stackPointer from ushort to short
Many issues were caused by changing memoryPointer and stackPointer to
ushort. These values would likely be better off as ushorts in the future
due to no need for negative values, but this change must be done when
the core is functioning as intended.

+++ Source/GameBeak-Frontend/Classes/Core.cs
@@ -81,7 +81,7 @@ namespace GameBeak_Frontend.Classes
                     //if (!cpu.checkForHaltOrInterrupt())
                     if (!Core.beakCPU.checkForHaltOrInterrupt())
                     {
-                        Core.beakCPU.selectOpcode(Core.beakMemory.readMemory(Core.beakMemory.memoryPointer++));
+                        Core.beakCPU.selectOpcode(Core.beakMemory.readMemory((ushort)Core.beakMemory.memoryPointer++));
                     }
                     else
                     {
